@@ -16,69 +16,33 @@
   "author": "CRIMX<straybugs@gmail.com>",
   "license": "MIT",
   "scripts": {
-    "postinstall": "lerna bootstrap && yarn build",
-    "build": "lerna run --scope 'observable-*' build && lerna bootstrap",
-    "test": "jest --coverage",
-    "docs:dev": "vuepress dev docs",
-    "docs:build": "vuepress build docs",
-    "commit": "git-cz",
-    "lint": "eslint packages/**",
-    "release": "lerna version --conventional-commits",
-    "publish": "lerna publish from-package",
-    "ship": "lerna run build && yarn run publish"
-  },
-  "husky": {
-    "hooks": {
-      "commit-msg": "commitlint -E HUSKY_GIT_PARAMS"
-    }
+    "build": "pnpm -r build",
+    "test": "pnpm -F observable-hooks test",
+    "docs:dev": "NODE_OPTIONS=--openssl-legacy-provider vuepress dev docs",
+    "docs:build": "NODE_OPTIONS=--openssl-legacy-provider vuepress build docs",
+    "lint": "eslint --ext .ts,.tsx,.js,.mjs . && prettier --check .",
+    "lint:fix": "eslint --ext .ts,.tsx,.js,.mjs . --fix && prettier -w .",
+    "release": "standard-version"
   },
   "devDependencies": {
-    "@babel/preset-react": "^7.14.5",
-    "@babel/preset-typescript": "^7.14.5",
-    "@commitlint/cli": "^8.1.0",
-    "@commitlint/config-conventional": "^8.1.0",
-    "@testing-library/react-hooks": "^3.2.1",
-    "@types/jest": "^26.0.23",
-    "@types/react": "^16.9.19",
-    "@types/react-dom": "^16.9.5",
-    "@types/react-test-renderer": "^16.8.2",
-    "@types/scheduler": "^0.16.1",
-    "@types/use-subscription": "^1.0.0",
-    "@typescript-eslint/eslint-plugin": "^4.15.2",
-    "@typescript-eslint/parser": "^4.15.2",
-    "@vuepress/plugin-pwa": "^1.2.0",
-    "commitizen": "^4.0.3",
-    "coveralls": "^3.0.5",
-    "cross-env": "^5.2.0",
-    "cz-conventional-changelog": "^3.0.2",
-    "eslint": "^7.28.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-config-standard": "^16.0.3",
-    "eslint-plugin-import": "^2.23.4",
-    "eslint-plugin-jest": "^24.3.6",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-prettier": "^3.4.0",
-    "eslint-plugin-promise": "^5.1.0",
-    "eslint-plugin-react": "^7.24.0",
-    "eslint-plugin-standard": "^5.0.0",
-    "experimental_react": "npm:react@experimental",
-    "experimental_react-dom": "npm:react-dom@experimental",
-    "experimental_react-test-renderer": "npm:react-test-renderer@experimental",
+    "@types/node": "^20.1.7",
+    "@typescript-eslint/eslint-plugin": "^6.3.0",
+    "@typescript-eslint/parser": "^6.3.0",
+    "@vuepress/plugin-pwa": "^1.9.10",
+    "cross-env": "^7.0.3",
+    "eslint": "^8.47.0",
+    "eslint-config-prettier": "^9.0.0",
     "fs-extra": "^10.0.0",
-    "husky": "^3.0.1",
-    "jest": "^27.0.4",
-    "lerna": "^3.20.2",
-    "prettier": "^2.3.1",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-test-renderer": "^16.13.1",
-    "rxjs": "^6.6.7",
-    "shx": "^0.3.3",
-    "standard-version": "^8.0.1",
-    "ts-jest": "^27.0.3",
-    "typescript": "~4.1.2",
-    "use-subscription": "^1.4.1",
-    "vuepress": "^1.2.0",
-    "vuepress-plugin-container": "^2.1.2"
+    "prettier": "^2.8.4",
+    "vuepress": "^1.9.10",
+    "vuepress-plugin-container": "^2.1.5"
+  },
+  "peerDependencyRules": {
+    "ignoreMissing": [
+      "eslint",
+      "react",
+      "react-dom",
+      "typescript"
+    ]
   }
 }
