@@ -1,39 +1,39 @@
-import { useObservableState, useObservable } from 'observable-hooks'
-import * as React from 'react'
+import { useObservableState, useObservable } from "observable-hooks";
+import * as React from "react";
 import {
   map,
   distinctUntilChanged,
   switchMap,
   catchError,
   startWith,
-  tap
-} from 'rxjs/operators'
-import { of, timer, Observable } from 'rxjs'
+  tap,
+} from "rxjs/operators";
+import { of, timer, Observable } from "rxjs";
 
 export interface SuggestsItem {
-  href: string
-  title: string
-  content: string
+  href: string;
+  title: string;
+  content: string;
 }
 
-export type SuggestsList = SuggestsItem[]
+export type SuggestsList = SuggestsItem[];
 
-export type SuggestsFetcher = (text: string) => Observable<SuggestsList>
+export type SuggestsFetcher = (text: string) => Observable<SuggestsList>;
 
 export interface SuggestsProps {
-  text: string
-  fetchFunc: SuggestsFetcher
+  text: string;
+  fetchFunc: SuggestsFetcher;
 }
 
-const StateDefault: React.FC = () => null
+const StateDefault: React.FC = () => null;
 
 const StateLoading: React.FC = () => (
   <div className="notification is-primary has-text-centered">Searching...</div>
-)
+);
 
 const StateFinish: React.FC<{ list: SuggestsList }> = props => (
   <ul>
-    {props.list!.map(item => (
+    {props.list.map(item => (
       <li className="box" key={item.href}>
         <strong>
           <a href={item.href} target="_blank" rel="noopener noreferrer">
@@ -44,13 +44,13 @@ const StateFinish: React.FC<{ list: SuggestsList }> = props => (
       </li>
     ))}
   </ul>
-)
+);
 
 const StateError: React.FC = () => (
   <div className="notification is-danger has-text-centered">
     Failed fetching...
   </div>
-)
+);
 
 /** Reusable Suggests Component */
 export const Suggests: React.FC<SuggestsProps> = props => {
@@ -65,7 +65,7 @@ export const Suggests: React.FC<SuggestsProps> = props => {
               // searching state quickly. But no actual request
               // is performed until the delay is hit.
               timer(750).pipe(
-                tap(() => console.log('>>> really start searching...')),
+                tap(() => console.log(">>> really start searching...")),
                 switchMap(() => fetchFunc(text)),
                 map(suggests => <StateFinish list={suggests} />),
                 // handle errors on sub-stream so that main stream stays alive
@@ -78,6 +78,6 @@ export const Suggests: React.FC<SuggestsProps> = props => {
         )
       ),
     [props.text, props.fetchFunc]
-  )
-  return useObservableState(status$, () => <StateDefault />)
-}
+  );
+  return useObservableState(status$, () => <StateDefault />);
+};
