@@ -1,10 +1,10 @@
-import { useObservableState, useObservable } from 'observable-hooks'
-import * as React from 'react'
-import { map, switchMap, startWith } from 'rxjs/operators'
-import { interval } from 'rxjs'
+import { useObservableState, useObservable } from "observable-hooks";
+import * as React from "react";
+import { map, switchMap, startWith } from "rxjs/operators";
+import { interval } from "rxjs";
 
 export interface InputTimeAgoProps {
-  text: string
+  text: string;
 }
 
 export const InputTimeAgo: React.FC<InputTimeAgoProps> = props => {
@@ -22,25 +22,25 @@ export const InputTimeAgo: React.FC<InputTimeAgoProps> = props => {
       [props.text]
     ),
     () => fromNow(0)
-  )
+  );
 
   return (
     <div className="notification">
-      You typed {props.text ? `"${props.text}"` : 'nothing'} {secondsAgo}.
+      You typed {props.text ? `"${props.text}"` : "nothing"} {secondsAgo}.
     </div>
-  )
-}
+  );
+};
 
 function fromNow(diff: number): string {
-  const minute = (diff / 60) | 0
-  const second = diff - minute * 60
+  const minute = (diff / 60) | 0;
+  const second = diff - minute * 60;
   const mstr =
-    minute < 1 ? '' : minute === 1 ? `a minute and ` : `${minute} minutes and `
+    minute < 1 ? "" : minute === 1 ? `a minute and ` : `${minute} minutes and `;
   const sstr =
     second < 1
-      ? 'just now'
+      ? "just now"
       : second === 1
       ? `a second ago`
-      : `${second} seconds ago`
-  return mstr + sstr
+      : `${second} seconds ago`;
+  return mstr + sstr;
 }
