@@ -1,22 +1,22 @@
 import {
   useObservableCallback,
   useSubscription,
-  pluckCurrentTargetValue
-} from 'observable-hooks'
-import * as React from 'react'
+  pluckCurrentTargetValue,
+} from "observable-hooks";
+import * as React from "react";
 
 export interface InputProps {
-  text: string
-  onChange: (text: string) => any
+  text: string;
+  onChange: (text: string) => any;
 }
 
 export const CustomInput: React.FC<InputProps> = props => {
   const [onChange, textChange$] = useObservableCallback<
     string,
     React.FormEvent<HTMLInputElement>
-  >(pluckCurrentTargetValue)
+  >(pluckCurrentTargetValue);
 
-  useSubscription(textChange$, props.onChange)
+  useSubscription(textChange$, props.onChange);
 
   return (
     <input
@@ -26,5 +26,5 @@ export const CustomInput: React.FC<InputProps> = props => {
       value={props.text}
       onChange={onChange}
     />
-  )
-}
+  );
+};
