@@ -1,8 +1,8 @@
-import { identity } from '../src'
+import { identity } from "../src";
 
-describe('identity', () => {
-  it('should return the first argument given', () => {
-    const object = { name: 'CRIMX' }
-    expect(identity(object)).toBe(object)
-  })
-})
+describe("identity", () => {
+  it("should return the first argument given", () => {
+    const object = { name: "CRIMX" };
+    expect(identity(object)).toBe(object);
+  });
+});
