@@ -1,8 +1,8 @@
-import { Observable, BehaviorSubject, Subject, Subscription } from 'rxjs'
+import { BehaviorSubject, Observable, Subject, Subscription } from "rxjs";
 
 interface Handler<T = any> {
-  suspender: Promise<T>
-  resolve: (value?: T) => void
+  suspender: Promise<T>;
+  resolve: (value?: T) => void;
 }
 
 /**
@@ -13,29 +13,27 @@ export class ObservableResource<TInput, TOutput extends TInput = TInput> {
    * Unlike Promise, Observable is a multiple push mechanism.
    * Only force update when Suspense needs to restart.
    */
-  readonly shouldUpdate$$ = new Subject<true>()
+  readonly shouldUpdate$$ = new Subject<true>();
 
   get isDestroyed(): boolean {
-    return this._isDestroyed
+    return this._isDestroyed;
   }
 
-  private handler: Handler | null = this.getHandler()
+  private handler: Handler | null = this.getHandler();
 
   public valueRef$$ = new BehaviorSubject<{ current: TOutput } | undefined>(
     undefined
-  )
+  );
 
-  private error: any = null
+  private error: unknown = null;
 
-  public input$: Observable<TInput>
+  public input$: Observable<TInput>;
 
-  private subscription: Subscription
+  private subscription: Subscription;
 
-  private isSuccess = (value: TInput): value is TOutput => true
+  private isSuccess = (_value: TInput): _value is TOutput => true;
 
-  private _isDestroyed = false
-
-  private _updateCount = 0
+  private _isDestroyed = false;
 
   /**
    * @param input$ An Observable.
@@ -50,114 +48,115 @@ export class ObservableResource<TInput, TOutput extends TInput = TInput> {
       : (value: TInput) => value is TOutput
   ) {
     if (isSuccess) {
-      this.isSuccess = isSuccess as (value: TInput) => value is TOutput
+      this.isSuccess = isSuccess as (value: TInput) => value is TOutput;
     }
 
-    this.input$ = input$
+    this.input$ = input$;
 
     this.subscription = input$.subscribe({
       next: this.handleNext,
       error: this.handleError,
-      complete: this.handleComplete
-    })
+      complete: this.handleComplete,
+    });
   }
 
   read(): TOutput {
     if (this.error) {
-      throw this.error
+      throw this.error;
     }
     if (this.handler) {
-      throw this.handler.suspender
+      throw this.handler.suspender;
     }
-    return this.valueRef$$.value?.current!
+    // eslint-disable-next-line @typescript-eslint/no-non-null-asserted-optional-chain
+    return this.valueRef$$.value?.current!;
   }
 
   reload(newInput$?: Observable<TInput>): void {
     if (this._isDestroyed) {
-      throw new Error('Cannot reload a destroyed Observable Resource')
+      throw new Error("Cannot reload a destroyed Observable Resource");
     }
 
     if (newInput$) {
-      this.input$ = newInput$
+      this.input$ = newInput$;
     }
 
-    this.subscription.unsubscribe()
+    this.subscription.unsubscribe();
 
-    this.error = null
+    this.error = null;
 
     if (this.handler) {
-      this.handler.resolve()
-      this.handler = this.getHandler()
+      this.handler.resolve();
+      this.handler = this.getHandler();
     }
 
     this.subscription = this.input$.subscribe({
       next: this.handleNext,
       error: this.handleError,
-      complete: this.handleComplete
-    })
+      complete: this.handleComplete,
+    });
   }
 
   destroy(): void {
-    this._isDestroyed = true
-    this.subscription.unsubscribe()
-    this.shouldUpdate$$.complete()
+    this._isDestroyed = true;
+    this.subscription.unsubscribe();
+    this.shouldUpdate$$.complete();
     if (this.handler) {
-      this.error = new Error('Resource has been destroyed.')
-      const { resolve } = this.handler
-      this.handler = null
-      resolve()
+      this.error = new Error("Resource has been destroyed.");
+      const { resolve } = this.handler;
+      this.handler = null;
+      resolve();
     }
   }
 
   private handleNext = (value: TInput): void => {
-    this.error = null
+    this.error = null;
     if (this.isSuccess(value)) {
       if (this.valueRef$$.value?.current !== value) {
-        this.valueRef$$.next({ current: value })
+        this.valueRef$$.next({ current: value });
       }
       if (this.handler) {
         // This will also remove the initial
         // suspender if sync values are emitted.
-        const { resolve } = this.handler
-        this.handler = null
-        resolve()
+        const { resolve } = this.handler;
+        this.handler = null;
+        resolve();
       }
     } else if (!this.handler) {
       // start a new Suspense
-      this.handler = this.getHandler()
-      this.shouldUpdate$$.next(true)
+      this.handler = this.getHandler();
+      this.shouldUpdate$$.next(true);
     }
-  }
+  };
 
-  private handleError = (error: any): void => {
-    this.error = error
+  private handleError = (error: unknown): void => {
+    this.error = error;
     if (this.handler) {
-      const { resolve } = this.handler
-      this.handler = null
+      const { resolve } = this.handler;
+      this.handler = null;
       // Errors thrown from the request is not catchable by error boundaries.
       // Here we resolve the suspender and let this.read throw the error.
-      resolve()
+      resolve();
     } else {
-      this.shouldUpdate$$.next(true)
+      this.shouldUpdate$$.next(true);
     }
-  }
+  };
 
   private handleComplete = (): void => {
     if (this.handler) {
-      this.error = new Error('Suspender ended unexpectedly.')
-      const { resolve } = this.handler
-      this.handler = null
+      this.error = new Error("Suspender ended unexpectedly.");
+      const { resolve } = this.handler;
+      this.handler = null;
       // Errors thrown from the request is not catchable by error boundaries.
       // Here we resolve the suspender and let this.read throw the error.
-      resolve()
+      resolve();
     }
-  }
+  };
 
   private getHandler(): Handler {
-    const handler: Partial<Handler> = {}
+    const handler: Partial<Handler> = {};
     handler.suspender = new Promise(resolve => {
-      handler.resolve = resolve
-    })
-    return handler as Handler
+      handler.resolve = resolve;
+    });
+    return handler as Handler;
   }
 }
