@@ -1,6 +1,6 @@
-import type { MutableRefObject, RefObject } from 'react'
-import { BehaviorSubject } from 'rxjs'
-import { useState } from 'react'
+import type { MutableRefObject, RefObject } from "react";
+import { BehaviorSubject } from "rxjs";
+import { useState } from "react";
 
 /**
  * Returns a mutable ref object and a BehaviorSubject.
@@ -11,7 +11,7 @@ import { useState } from 'react'
  */
 export function useObservableRef<TValue>(
   initialValue: TValue
-): [MutableRefObject<TValue>, BehaviorSubject<TValue>]
+): [MutableRefObject<TValue>, BehaviorSubject<TValue>];
 /**
  * Returns a ref object and a BehaviorSubject.
  *
@@ -21,7 +21,7 @@ export function useObservableRef<TValue>(
  */
 export function useObservableRef<TValue>(
   initialValue: TValue | null
-): [RefObject<TValue>, BehaviorSubject<TValue>]
+): [RefObject<TValue>, BehaviorSubject<TValue>];
 /**
  * Returns a mutable ref object and a BehaviorSubject.
  *
@@ -31,18 +31,18 @@ export function useObservableRef<TValue>(
  */
 export function useObservableRef<TValue = undefined>(
   initialValue?: TValue
-): [MutableRefObject<TValue | undefined>, BehaviorSubject<TValue | undefined>]
+): [MutableRefObject<TValue | undefined>, BehaviorSubject<TValue | undefined>];
 export function useObservableRef<TValue>(
   initialValue?: TValue
 ): [MutableRefObject<TValue | undefined>, BehaviorSubject<TValue | undefined>] {
-  const [value$] = useState(() => new BehaviorSubject(initialValue))
+  const [value$] = useState(() => new BehaviorSubject(initialValue));
   const [ref] = useState<MutableRefObject<TValue | undefined>>(() => ({
     get current(): TValue | undefined {
-      return value$.value
+      return value$.value;
     },
     set current(value: TValue | undefined) {
-      value$.next(value)
-    }
-  }))
-  return [ref, value$]
+      value$.next(value);
+    },
+  }));
+  return [ref, value$];
 }
